Fix error for state with pure hydrogen
When trying to instantiate a state with pure hydrogen we would get the
following error:
ValueError: Could not load these fluids: HYDROGEN|HYDROGEN

This happened because of the workaround implemented to be able to plot
a phase diagram of a pure substance (see fbe41d0), which would create
a mixture with the same component. For some reason this worked for most
of the fluids, but it would raise this error for hydrogen.
The modification implemented here will make a mixture with another
fluid to avoid this error.