Change state.fluid dict to reflect pure fluid
In this function we have a workaround to enable building phase
envelopes for pure substances. In the first part of this workaround we
try to use a composition that is actually the same fluid duplicated
with molar fractions of 1 - 1e-15 and 1e-15. If this works, at the end
we would have a fluid dictionary with only the last molar fraction,
since the key value in a dictionary can not be duplicated:
>>> methane = ccp.State.define(p=100000, T=300, fluid='methane')
>>> methane.fluid
{'METHANE': 1e-15}

To solve this we force the value to 1.