Remove feature to accept pure fluid as string
The pure fluid capability was actually a hack. Since we want to be able
to plot the phase envelope for all states and we had an issue to do
that with pure fluid (see CoolProp issue #1544), we were actually
changing the pure fluid to a mixture where we would have something like
{pure_fluid: 1 - 1e-15, another_component: 1e-15}. So it was not
actually a pure fluid and I think it is better to make it explicit that
we only work with mixtures and if the user wants the can use the
solution above with an additional component with mol frac of 1e-15.