Add b, D and similarity attributes
Added the impeller b and D as optional arguments, in case they are not
given, their value will be set when they are associated with an
impeller. Also added the comparison between similarity numbers
(phi_ratio, psi_ratio, reynolds_ratio and mach_diff). These are set to
1. (or 0) when the point is instantiated with the __init__ constructor.
A second constructor called convert_from will be added and in this case
these values will be calculated by comparing the original and converted
point.