Change Point class api
This changes the Point class api. Now it is possible to call the
constructor with non dimensional arguments such as phi and psi. The
__init__ method will evaluate which arguments are provided and it will
call an specific method to calculate the point attributes that were not
given in the constructor call.
This also removes several methods from the Point class making them
functions. The purpose is to make the code more simple to maintain.