Remove methods that are now functions
The Point class has been modified and now it has the b and D
attributes. These methods are now all available as functions in the
points module, so there is no need to have them in the Impeller class.
The idea is that the Impeller will work only as an object to group
points.
Conversions and plots will always call the methods that are available
in the point objects.