Fix pint quantities pickle
Since multiprocess requires pickle to be working we have to modify how
we create the pint unit registry so that the code uses the same
registry throughout the code. As explained in pint's tutorial docs:

"If you are pickling and unplicking Quantities within your project, you
should also define the registry as the application registry".