Move imp_fd calculation to name==main part
This is necessary to allow the parallel computation that is carried out
within the impeller to be done in windows machines. In windows the
multiprocessing module uses 'spawn' instead of 'fork' to start children
processes. In this case the main module is imported and we have to avoid
that this part of the code is executed again by the children process,
since this would spawn more process repeating the cycle and leading to a
RunTimeError.