Change how we calculate the last curve point
In the flow_v_range for loop we only consider points that have a head
higher than the minimum value calculated by the interpolation method
implemented in the get_interpolated_value function. Errors in the loop
are now discarded. These errors could happen especially for point with
higher flow, where the curve interpolation function would cause the
head to be very low, causing errors in the state convergence.
After going through the loop we now add a final point that finds the
flow_v for the min_head calculated. For this we use the brentq function,
since we know that the flow we are searching is between surge and choke
(flow_v_range[0] and flow_v_range[-1]). The newton method would fail in
some cases.
The brentq method fails when f(a) and f(b) have the same signal, but in
this case we will already have the minimum head when going through the
for loop (flow_v_range[-1]).