Fix how converted points are calculated
After calculating the speed mean for the converted points in a curve,
each point was instantiated considering that new speed. The problem
is that points were being instantiated directly with Point(), and this
approach would ignore that the points were being converted, therefore,
attributes such as volume_ratio_ratio would return a value of 1., which
is not true since we are changing the speed for which the point have
been previously calculated.