Change how min and max flow points are selected
This changes how the min and max flow points are selected when importing
curves from csv files.

The previous code would fail when we had a max flow point defined by
the head curve. When we tried to calculate the efficiency for that flow,
since it is normal to have a big drop in efficiency near the max flow,
we could end up with a very low efficiency (e.g. 20~30%), and this would
cause the code to break with an error in the state convergence.