Add check units from ROSS
We are now using the improved version of check units from ROSS. In this
case we need to do some modifications due to issues with variables such
as G_s (the s will be checked as entropy after splitting on '_'). On
solution to this would be to look up the complete name first in the
unit dictionary, if it is there we don't go to the split on '_' step.