Add warning for interpolation error
Added a comparison between head and eff curves interpolated with
interp1d cubic and UnivariateSpline. In cases where they are too
different a warning will be raised so that the user can evaluate if
an error is present in the interpolation process, maybe due to points
that would not produce a smooth curve.