Improve normalization
This code tries to guarantee that sum(molar_fractions) is always equal
to 1.0 and that the fractions for the normalized components are always
the same no matter what the dictionary order is. This is achieved by
using sort in some parts of the code to avoid float point erros (sum)
and keep repeatability.