Add module for compressor tests
This module will have the StraightThrough and the BackToBack
compressors implemented. The user will need the guarantee point and the
test points to instantiate the class.
An additional Point class will be created for each case to consider
other parameters related to the point during the test, such as balance
line flow, oil flow, etc.