Change imp_test to just list of points
We cannot create an impeller with only one point per speed since we
perform curve interpolation and other tasks that require more than one
point for each curve. The simple solution for now was to create a list
of point for the test and the impeller will only be created to the
converted condition where all points will have the same speed.