Add point calculation to back_to_back
The point calculation will calculate a point based on flange conditions,
but the power returned by this point is the 'real' power consumed by the
rotor.
This is accomplished by calculating the point from the Impeller
generated with the flange points and then replacing the Point.power
attribute with the calculated power based on the rotor conditions.