Add possibility to give mbal to only 1 point
This adds to calculate tests where mbal is not measured in all points.
In this case the k_mean for the end seal is calculated and is used to
get the end seal flow for the other points. It should also be noted that
the user should not pass mbal as zero, since this value will impact the
calculation of the k_mean.