Add possibility to give md1f to only 1 point
This adds to calculate tests where md1f is not measured in all points.
In this case the k_mean for the div wall is calculated and is used to
get the div wall flow for the other points. It should also be noted that
the user should not pass md1f as zero, since this value will impact the
calculation of the k_mean.