Remove seal gas temp from second sect calculation
The seal gas temperature is not actually needed for the second section
calculation, since it does not have an impact in the rotor suction
temperature. The seal gas flow only has an influence on the calculation
of how much mass flow is being removed from the second section rotor
suction and being sent to the first section rotor suction.