Change 2 sect points from Point2Sec to Point1Sec
These names are a bit confusing. Although we use the class Point2Sec
for the first section of a compressor with 2 sections (back-to-back),
we use the Point1Sec class for the points in the second section. This
happens because for the second section we do not need the division wall
calculation, since this section is not affected by it. So, we can use
the more simple Point1Sec class in this case.