Change fill values for div and m1df
If we only give some values for m1df, after running, the code would fill
the values for the mdiv. This would cause problems when we would try to
run the spreadsheet again, since now the mdiv cells have some values.
In general I think we should avoid filling values where the user should
make inputs, with the risk of having bugs and some inconsistent
behavior.