Fix cases where leakages are set to 'No'
Cases where the user does not want to take into account the leakages
need to have values set to zero instead of None. These 0 values are only
used within the code, and we do not fill the spreadsheet with values to
avoid errors when the user has to run the code again. This way,
everytime the user runs the code the inputs are the same.