Add macro to find ccp
We have added a macro to run at the file startup with the purpose of
checking if python is available and after that finding the correct
location of the ccp installation. With this, it is now possible to move
the spreadsheets to different folder without the need to copy the script
to the same folder.