Create __new__ method to enable State(...)
Adds a __new__ method to deal with the subclassing of a cython class.
This is explained here: http://stackoverflow.com/questions/18260095/

Basically, before __init__ is called there is a call to __new__ and
then to __cinit__. If we do not implement __new__, the call to __cinit__
will have the wrong arguments and will fail.

A deprecation warning has also been added to the State.define() method.