Add attributes to store values given
If div_wall, first_discharge or balance_line flow are not given they
are calculated based on the seal k calculated from the other points.
The calculated values are then assigned to the attributes. If we create
another compressor object with the same points, the code will not run
in the same way if we check for these values and instead of None they
now have calculated values. This would actually result in negative
values for the div_wall flow after running some iterations.
Therefore, to guarantee that the code will always give the same results,
we store the original values used in instantiation in protected
attributes.