Move to one-page app
For now multi-page apps have an issue that the widget are not persistent
so when the user goes from one page to another and then back to the
first page, widgets such as text_input will be cleared and the user will
have to type the values again.
It seems that this will change soon as we can see in this discussion:
https://discuss.streamlit.io/t/keyed-widget-state-persistence-discussion-possible-fixes/37359/6
The issue #6074 in the streamlit repository is also tracking this, and
it seems that this behavior will change soon.