Move session name to inside the form
This avoids unnecessary rerunning the code due to a change in the
session name, which was actually causing some widgets within the web
page to get their value erased for some reason that I was not able to
identify.

I have also changed the clear_on_submit to False, because this was also
causing a strange behavior where I was not able to change the session
name and save another file. For some reason the name would always stay
the same.

I have also moved the save button to inside the form, so that it can get
the correct session name to write the saved file. Otherwise, it would
use the old session name instead.