Add option to keep same speed as original impeller
When converting an impeller to a different suction condition, the value
of volume_ratio_ratio is kept constant and a new speed is found. This
speed would have a different value from the speed in the original curve.
This is not always what the user wants, since some curves would even go
out of the compressor speed range.
With the speed='same' option, the impeller is converted and them new
curves are generated by interpolating the converted curves to the speed
values of the original impeller.