Implement new tests for FlowOrifice
The new tests will verify if changing to equivalent values/units will
return the same flow mass value, so we can evaluate if units conversion
is working as expected.

We have also added a test to check if FlowOrifice works with floats,
assuming that the floats are provided already at the default SI units.