Fix filter error when data type is not provided
The for loop has been modified to iterate over the data type dict
instead of the dataframe columns. This way it will not result in an
error if the user does not provide a data type for all columns of the
dataframe.
We have also implemented a speed filter in the function to remove rows
where the equipment is not in operation.