Add common file to store some parameters
The common module can be used to store data/parameters that are shared
between the ccp_app_straight_through.py and ccp_app_back_to_back.py.
This way we avoid having to modify these parameters in both files
everytime we need to update something.
In the future we could add more to this file.