Add code to force expansion when loading file
We now force the expansion of the expanders when a file is loaded. This
fixes a behavior where a file is loaded, the user presses the calculate
button and the data inside the expanders that were collapsed get lost.
Since we do not lose the data if each expander is open before we press
the calculate button, we have decided to force this when loading the
file to solve the issue.