Add clusters based on ps, Ts and speed_sound
Add cluster based on ps, Ts and speed_sound so that we can have a better
result when comparing operational point with converted curves.
The more 'correct' of way to compare the operational point, would be to
convert the design curves to that suction condition and then perform the
comparison. This becomes impractical when we have several data points,
since the conversion process is computationally intensive.
Therefore, the separation of points in clusters and conversion of curves
for each cluster is a compromise between time for calculation and
precision in the results.