Change interpolation method inside load_from_dict
The use of interp1d with kind=3 was causing the efficiency interpolation
to overshoot in some cases, returning interpolated values that would
not make sense for the efficiency and would cause the point calculation
to not converge.
To remove the overshoot problem we need to use a monotone cubic
interpolant. The available options were the Pchip and Akima
interpolators. We decided on the Pchip interpolator due to its
capability of extrapolating, which is necessary since we get the min
and max value of both provided curves and these values will have to be
extrapolated in one of the interpolations.