Add option to pass a list of impellers in convert
This option enables passing a list of impellers to the .convert_from
method. The code will then evaluate which impeller is best suited to be
used in the conversion process based on the suction speed of sound of
each impeller compared to the speed of sound of the new suction.