Fix surge line
The list of attributes was a list of lists, making it impossible to plot
the surge line.
The additional point for the specified speed has also been removed from
the surge line points, since we would need to sort the surge points
before plotting, and we would not gain anything from adding this point,
since it is already an interpolation from the other points.