Change criterion for diverging efficiency
For the first iteration to find the discharge state we assume an
isentropic compression and in the initial steps it is possible that we
get efficiencies higher than 1.
We are now setting the ceiling to 1.5, but we should evaluate if for
some cases (e.g. first section of back to back compressors) it is not
reasonable to get higher efficiencies.