Fix expected point calculation
In the evaluation module, to calculate the expected point we were using
the flow_m value. This would cause an error in the final calculated
value since flow_m would be converted to flow_v using the 'new' impeller
suction pressure and temperature, which are normally different from the
operational values. This happens because we are not actually converting
the curves for each and every operational point, instead we convert only
for a determined number of 'cluster center' points to decrease the time
necessary to compute everything.