Move plot generation to function with cache
To avoid streamlit regenerating the plots after each user interaction we
have moved the plot generation to inside a function with the
st.cache_data decorator.

To enable the use of cache the objects have to
be hashable/serialazable. We have implemented __hash__ for the
ccp.Impeller and the object is also pickable, but for some reason
streamlit could not serialize it.

To solve the problem we have decided to name the impeller argument as
_impeller, this ignores the argument in the creation of the cache.

To keep the cache working we had to implement a manual invalidation of
the cache by passing an additional argument with hash(impeller).